Add room name to BoardCard and fix navigation bug
When going to a room and exiting, it would redirect to the home page. This happened likely because of a race condition across two useEffects. One useEffect changes the state of the selectedRoom, and the other useEffect used the selectedRoom to check if the room existed. Now the second useEffect, will only run when roomId from useParams is not undefined, which is what we want.